+++ packages/terra-framework-docs/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## 1.27.0 - (July 25, 2023)
+
+* Changed
+  * Minor dependency version bump
+
 ## 1.26.0 - (July 20, 2023)
 
 * Added
@@ -15,6 +20,7 @@
   * Updated the path name for the examples in doc site for `terra-embedded-content-consumer` to resolve missing files.
   * Minor formatting changes in the `uuid` testing section for several packages.
   * Updates tests amd examples for `terra-worklist-data-grid` to show resize capability.
+  * Updated examples for `terra-tabs` with conflicting ids.
 
 * Added
   * Added alpha release disclaimer for `terra-worklist-data-grid`.